requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-guardian==2.4.0
django-leaflet==0.29.0
django-location-field==2.7.1
django-map-widgets==0.4.1
django-modeltranslation==0.18.10
django-ranged-response==0.2.0
django-rest-passwordreset==1.3.0
django-rosetta==0.9.9
django-simple-captcha==0.5.17
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
Pillow==9.5.0
polib==1.2.0
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
